Add typings for saved_object_finder (#30067)
