+++ src/ui/public/registry/vis_types.d.ts
@@ -16,26 +16,10 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import { Direction } from '@elastic/eui/src/services/sort/sort_direction';
 
-// TODO: Remove once typescript definitions are in EUI
+import { VisType } from '../vis';
+import { UIRegistry } from './_registry';
 
-declare module '@elastic/eui' {
-  export const EuiWrappingPopover: React.SFC<any>;
-  export const EuiCopy: React.SFC<any>;
-  export const EuiOutsideClickDetector: React.SFC<any>;
-  export const EuiSideNav: React.SFC<any>;
-
-  export interface EuiTableCriteria {
-    page: { index: number; size: number };
-    sort?: {
-      field?: string;
-      direction?: Direction;
-    };
-  }
-  export const EuiBasicTable: React.ComponentClass<{
-    onTableChange?: (criteria: EuiTableCriteria) => void;
-    sorting: { sort?: EuiTableCriteria['sort'] };
-    [key: string]: any;
-  }>;
-}
+declare type VisTypesRegistryProvider = UIRegistry<VisType> & {
+  byName: { [typeName: string]: VisType };
+};
